+++ R/used_synonyms.R
@@ -3,20 +3,21 @@
 #' @title Retrieve synonyms or taxon concepts used in a data set
 #'
 #' @description
-#' Plots records are rather linked to plant names than plant taxon concepts.
-#' The function `used_synonyms()` provides a quick report about synonyms used in
-#' a data set (a [vegtable-class] object) and their respective accepted names.
+#' Plots records are rather linked to plant names than plant taxon concepts and
+#' `used_synonyms()` lists all synonyms linked to records in a [vegtable-class]
+#' object, including their respective accepted names.
 #'
-#' Additionally, not all taxon concepts included in the taxonomic list (slot
-#' **species**) may be recorded in the plot observations.
-#' In that case the function `used_concepts()` will optimize the size of the
-#' taxonomic list by discarding taxa that are not "in use".
-#' Alternatively parents or children of these taxa may be included in the output
-#' data set.
+#' On the other side, the function `used_concepts()` produces a subset of the
+#' taxonomic list embeded in the slot **species** including only taxonomic
+#' concepts linked to records in the slot **samples**.
 #'
 #' @param x A [vegtable-class] object.
-#' @param keep_children Argument passed to [taxlist::get_children()].
-#' @param keep_parents Argument passed to [taxlist::get_parents()].
+#' @param keep_children A logical argument indicating whether children of
+#'     selected taxa should be included in the output or not.
+#'     This argument passed to [get_children()].
+#' @param keep_parents A logical argument indicating whether parents of
+#'     selected taxa should be included in the output or not.
+#'     This argument passed to [get_parents()].
 #' @param ... Further arguments to be passed from or to another methods.
 #'
 #' @return
@@ -45,92 +46,83 @@
 #' ## Synonyms used in the Kenya_veg
 #' Synonyms <- used_synonyms(Kenya_veg)
 #' head(Synonyms)
+#'
+#' ## Subset species list to used concepts
+#' species <- used_concepts(Kenya_veg)
+#' Kenya_veg@species
+#' species
+#'
 #' @rdname used_synonyms
-#'
-#' @exportMethod used_synonyms
-#'
-setGeneric(
-  "used_synonyms",
-  function(x, ...) {
-    standardGeneric("used_synonyms")
-  }
-)
+#' @export
+used_synonyms <- function(x, ...) {
+  UseMethod("used_synonyms", x)
+}
 
 #' @rdname used_synonyms
-#'
 #' @aliases used_synonyms,vegtable-method
-#'
-setMethod(
-  "used_synonyms", signature(x = "vegtable"),
-  function(x, ...) {
-    SYN <- x@samples$TaxonUsageID[!x@samples$TaxonUsageID %in%
-      x@species@taxonRelations$AcceptedName]
-    SYN <- data.frame(SynonymID = unique(SYN), stringsAsFactors = FALSE)
-    SYN$Synonym <- x@species@taxonNames$TaxonName[match(
-      SYN$SynonymID,
+#' @method used_synonyms vegtable
+#' @export
+used_synonyms.vegtable <- function(x, ...) {
+  SYN <- x@samples$TaxonUsageID[!x@samples$TaxonUsageID %in%
+    x@species@taxonRelations$AcceptedName]
+  SYN <- data.frame(SynonymID = unique(SYN), stringsAsFactors = FALSE)
+  SYN$Synonym <- x@species@taxonNames$TaxonName[match(
+    SYN$SynonymID,
+    x@species@taxonNames$TaxonUsageID
+  )]
+  SYN$SynonymAuthor <- x@species@taxonNames$AuthorName[
+    match(SYN$SynonymID, x@species@taxonNames$TaxonUsageID)
+  ]
+  SYN$TaxonConceptID <- x@species@taxonNames$TaxonConceptID[
+    match(SYN$SynonymID, x@species@taxonNames$TaxonUsageID)
+  ]
+  SYN$AcceptedNameID <- x@species@taxonRelations$AcceptedName[
+    match(
+      SYN$TaxonConceptID,
+      x@species@taxonRelations$TaxonConceptID
+    )
+  ]
+  SYN$AcceptedName <- x@species@taxonNames$TaxonName[
+    match(
+      SYN$AcceptedNameID,
       x@species@taxonNames$TaxonUsageID
-    )]
-    SYN$SynonymAuthor <- x@species@taxonNames$AuthorName[
-      match(SYN$SynonymID, x@species@taxonNames$TaxonUsageID)
-    ]
-    SYN$TaxonConceptID <- x@species@taxonNames$TaxonConceptID[
-      match(SYN$SynonymID, x@species@taxonNames$TaxonUsageID)
-    ]
-    SYN$AcceptedNameID <- x@species@taxonRelations$AcceptedName[
-      match(
-        SYN$TaxonConceptID,
-        x@species@taxonRelations$TaxonConceptID
-      )
-    ]
-    SYN$AcceptedName <- x@species@taxonNames$TaxonName[
-      match(
-        SYN$AcceptedNameID,
-        x@species@taxonNames$TaxonUsageID
-      )
-    ]
-    SYN$AcceptedNameAuthor <- x@species@taxonNames$AuthorName[
-      match(
-        SYN$AcceptedNameID,
-        x@species@taxonNames$TaxonUsageID
-      )
-    ]
-    return(SYN)
-  }
-)
+    )
+  ]
+  SYN$AcceptedNameAuthor <- x@species@taxonNames$AuthorName[
+    match(
+      SYN$AcceptedNameID,
+      x@species@taxonNames$TaxonUsageID
+    )
+  ]
+  return(SYN)
+}
 
 #' @rdname used_synonyms
-#'
-#' @exportMethod used_concepts
-#'
-setGeneric(
-  "used_concepts",
-  function(x, ...) {
-    standardGeneric("used_concepts")
-  }
-)
+#' @export
+used_concepts <- function(x, ...) {
+  UseMethod("used_concepts", x)
+}
 
 #' @rdname used_synonyms
-#'
-#' @aliases used_concepts used_concepts,vegtable-method
-#'
-setMethod(
-  "used_concepts", signature(x = "vegtable"),
-  function(x, keep_children = FALSE, keep_parents = FALSE, ...) {
-    concepts <- unique(x@species@taxonNames$TaxonConceptID[
-      x@species@taxonNames$TaxonUsageID %in%
-        x@samples$TaxonUsageID
-    ])
-    z <- x@species
-    z@taxonRelations <- z@taxonRelations[
-      z@taxonRelations$TaxonConceptID %in% concepts,
-    ]
-    z <- clean(z)
-    if (keep_children) {
-      z <- get_children(x@species, z)
-    }
-    if (keep_parents) {
-      z <- get_parents(x@species, z)
-    }
-    return(z)
+#' @aliases used_concepts,vegtable-method
+#' @method used_concepts vegtable
+#' @export
+used_concepts.vegtable <- function(x, keep_children = FALSE,
+                                   keep_parents = FALSE, ...) {
+  concepts <- unique(x@species@taxonNames$TaxonConceptID[
+    x@species@taxonNames$TaxonUsageID %in%
+      x@samples$TaxonUsageID
+  ])
+  z <- x@species
+  z@taxonRelations <- z@taxonRelations[
+    z@taxonRelations$TaxonConceptID %in% concepts,
+  ]
+  z <- clean(z)
+  if (keep_children) {
+    z <- get_children(x@species, z)
   }
-)
+  if (keep_parents) {
+    z <- get_parents(x@species, z)
+  }
+  return(z)
+}