+++ R/veg_diversity.R
@@ -10,17 +10,18 @@
 #'
 #' This function calls [taxa2samples()] to derive taxa from taxon usage names in
 #' slot **samples** and multiple records of species in a single plot will be
-#' merged by [stats::aggregate()].
+#' merged by [stats::aggregate][aggregate()].
 #'
 #' The functions `shannon()`, `evenness()`, and `dominance()` calculate the
-#' diversity index of Shannon, the evenness, and the dominance, respectively.
+#' diversity index of Shannon, the evenness, and the dominance
+#' (*1 - evenness*), respectively.
 #' Dominance is the complementary value to evenness (i.e. `1 - evenness`).
 #'
 #' The function `simpson()` calculates the Simpson's index using the alternative
 #' for vegetation plot observations.
 #'
 #' The function `richness()` counts the number of taxa per plot and can be used
-#' as alternative to [vegtable::count_taxa()].
+#' as alternative to [vegtable::count_taxa][count_taxa()].
 #'
 #' @param x A numeric vector containing the abundance of single species.
 #' @param na.rm A logical value indicating whether NA values should be removed
@@ -44,8 +45,10 @@
 #'     included as variables in the slot **header** of the input object.
 #'     If `'in_header = TRUE'`, you may assign the result of the function to the
 #'     input object.
-#' @param ... Further arguments passed to `'FUN'` by [aggregate()] in
-#'     function `'veg_diversity()'`.
+#' @param ... Further arguments passed among methods. In `'evenness()'` and
+#'     `'dominance()'`, these arguments are passed to `'shannon()'`.
+#'     In `'veg_diversity()'`, these arguments are passed to [aggregate()]
+#'     (actually to `'FUN'`).
 #'
 #' @return
 #' Functions `shannon()`, `evenness()`, `dominance()`, `simpson()`, and
@@ -55,6 +58,14 @@
 #' per plot observation (option `'in_header = FALSE'`) or a [vegtable-class]
 #' object with the calculated values inserted in the slot **header**
 #' (option `'in_header = TRUE'`).
+#'
+#' @examples
+#' ## Compare Evenness with Shannon index
+#' Kenya_veg <- cover_trans(x = Kenya_veg, to = "cover")
+#' Kenya_veg <- veg_diversity(object = Kenya_veg, weight = "cover")
+#' Kenya_veg <- veg_diversity(object = Kenya_veg, weight = "cover", FUN = evenness)
+#'
+#' with(Kenya_veg@header, plot(shannon, evenness))
 #'
 #' @aliases shannon
 #' @export
@@ -106,7 +117,8 @@ veg_diversity <- function(object, ...) {
 #' @method veg_diversity vegtable
 #' @export
 veg_diversity.vegtable <- function(object, weight, FUN = shannon,
-                                   aggr_fun = mean, arg_fun = list(), var_name, in_header = TRUE,
+                                   aggr_fun = mean, arg_fun = list(), var_name,
+                                   in_header = TRUE,
                                    ...) {
   if (!weight %in% names(object@samples)) {
     stop(paste0(