+++ NEWS.md
@@ -5,6 +5,8 @@ vegtable 0.1.8
 
 * New function `veg_diversity()` and functions calculating diversity indices:
   `shannon()`, `evenness()`, `dominance()`, `simpson()`, and `richness()`.
+* New function `new_relation()` inserting new relations into `vegtable` objects.
+  This function is internally called by `veg_relation()`.
 
 ## Improvements
 
@@ -14,7 +16,12 @@ vegtable 0.1.8
 * Parameter `in_header` defined in several functions is set as
   `in_header = TRUE`.
 * Former method for 'aggregate()' is now defined in function 'veg_aggregate()'.
-
+* Deprecated functions: `match_names()`, `merge_taxa()`. These functions are now
+  exclusive for objects of class `taxlist`.
+* Method `names()` for `vegtable` objects will retrieve a list with all names
+  from the respective slots.
+* New coercion methods including functions `as()` and `as<-`.
+  
 ## Bug Fixes
 
 * An issue was solved for function `taxa2samples()` when setting an argument in